requirements.txt
aiofiles==23.2.1
aiohttp==3.8.4
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
appdirs==1.4.4
argcomplete==1.10.3
argilla==1.6.0
arrow==1.2.3
arxiv==1.4.4
arxiv2bib==1.0.8
asteroid-filterbanks==0.4.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
Babel==2.12.1
backoff==2.2.1
beautifulsoup4==4.8.2
bibtexparser==1.4.0
bio==1.5.9
biopython==1.81
biothings-client==0.3.0
black==23.3.0
blinker==1.7.0
blobfile==2.0.2
bs4==0.0.2
build==1.1.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cmudict==1.0.22
cn2an==0.5.22
colorama==0.4.6
colorlog==6.8.2
commonmark==0.9.1
compressed-rtf==1.0.6
contourpy==1.0.7
cryptography==40.0.2
customtkinter==5.2.2
cycler==0.11.0
Cython==0.29.34
darkdetect==0.8.0
dataclasses-json==0.5.7
decorator==5.1.1
Deprecated==1.2.13
discord.py==2.3.2
Distance==0.1.3
distlib==0.3.6
distro==1.9.0
docopt==0.6.2
docutils==0.18.1
docx2txt==0.8
ebcdic==1.1.1
einops==0.7.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
extract-msg==0.28.7
faiss-cpu==1.7.3
fake-useragent==1.1.3
fastapi==0.110.0
feedparser==6.0.10
ffmpy==0.3.2
filelock==3.12.0
Flask==3.0.2
fonttools==4.39.3
free-proxy==1.1.1
frozenlist==1.3.3
fsspec==2024.3.1
g2p-en==2.1.0
google-api-core==2.11.0
google-api-python-client==2.122.0
google-auth==2.18.0
google-auth-httplib2==0.2.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.0
gprofiler-official==1.0.0
gptcache==0.1.10
gradio==4.22.0
gradio_client==0.13.0
greenlet==2.0.2
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2020.1.16
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.21.4
hyperframe==6.0.1
HyperPyYAML==1.2.2
identify==2.5.22
idna==3.4
imagesize==1.4.1
IMAPClient==2.1.0
importlib_metadata==7.0.2
importlib_resources==6.3.2
inflect==7.0.0
isodate==0.6.1
isort==5.12.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.4
langchain==0.0.142
lazy_loader==0.2
librosa==0.9.2
lightning==2.2.1
lightning-utilities==0.11.0
llama-index==0.5.23.post1
llvmlite==0.39.1
loguru==0.7.2
lxml==4.9.2
Mako==1.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
monotonic==1.6
more-itertools==10.2.0
MouseInfo==0.1.3
mpmath==1.3.0
msg-parser==1.2.0
msgpack==1.0.5
multidict==6.0.4
mygene==3.2.2
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.7.0
num2words==0.5.13
numba==0.56.4
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
omegaconf==2.3.0
openai==1.6.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
optuna==3.6.0
orjson==3.9.15
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
patsy==0.5.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.9.0
pendulum==3.0.0
Pillow==9.4.0
platformdirs==3.2.0
pooch==1.6.0
pre-commit==3.2.2
primePy==1.3
proces==0.1.7
proto-plus==1.22.2
protobuf==4.23.0
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.0.1
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyenchant==3.2.2
PyGetWindow==0.0.9
Pygments==2.15.0
PyJWT==2.8.0
PyMsgBox==1.0.9
pyopenjtalk-dict==0.3.4.dev2
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.8.1
PyPDF2==3.0.1
pyperclip==1.8.2
pypinyin==0.51.0
pyproject_hooks==1.0.0
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pytchat==0.5.5
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.9
python-pptx==0.6.21
pytorch-lightning==2.2.1
pytorch-metric-learning==2.4.1
pytweening==1.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyworld==0.3.4
pyworld-prebuilt==0.3.4.1
PyYAML==6.0
referencing==0.34.0
regex==2023.3.22
reportlab==3.6.12
requests==2.28.2
requests-file==2.0.0
requests-toolbelt==1.0.0
resampy==0.4.3
rfc3986==1.5.0
rich==13.0.1
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.3.3
safetensors==0.4.2
scholarly==1.7.11
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
selenium==4.8.3
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
simple-salesforce==1.12.5
six==1.12.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.4
speechbrain==1.0.0
SpeechRecognition==3.8.1
Sphinx==6.1.3
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.47
starlette==0.36.3
statsmodels==0.14.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboardX==2.6.2.2
textract==1.6.5
threadpoolctl==3.1.0
tiktoken==0.3.3
time-machine==2.14.0
tkinterweb==3.23.5
tokenizers==0.15.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.2.1
torchmetrics==1.3.2
tqdm==4.65.0
transformers==4.38.2
trio==0.22.0
trio-websocket==0.10.2
typer==0.9.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3
unstructured==0.6.2
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.28.1
virtualenv==20.22.0
Wand==0.6.11
webrtcvad==2.0.10
websockets==11.0.3
Werkzeug==3.0.1
win32-setctime==1.1.0
wordnet==0.0.1b2
wrapt==1.14.1
wsproto==1.2.0
xlrd==1.2.0
XlsxWriter==3.1.0
yarl==1.8.2
zeep==4.2.1
zipp==3.18.1
